Extract code from activity to fragment
