+++ app/src/main/res/layout/activity_element_form.xml
@@ -35,6 +35,7 @@
                 android:layout_marginEnd="8dp"
                 android:layout_marginStart="8dp"
                 android:layout_marginTop="8dp"
+                android:visibility="gone"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintHorizontal_bias="0.0"
                 app:layout_constraintStart_toStartOf="parent"
@@ -58,6 +59,7 @@
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintHorizontal_bias="0.0"
                 app:layout_constraintStart_toStartOf="parent"
+                android:visibility="gone"
                 app:layout_constraintTop_toBottomOf="@+id/element_form_title_input_layout">
 
                 <android.support.design.widget.TextInputEditText
@@ -164,7 +166,7 @@
                 android:layout_marginStart="8dp"
                 android:layout_marginTop="8dp"
                 app:gv_level="@{element.level}"
-                app:gv_scalesCount="10"
+                app:gv_scalesCount="5000"
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toBottomOf="@+id/imageView2" />
