+++ app/src/main/res/values-w600dp/config.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <item name="hasTwoPanes" type="bool">true</item>
+</resources>
