+++ app/src/main/res/values-w600dp/dimens.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <integer name="element_columns">2</integer>
+    <bool name="bool_has_two_panes">true</bool>
+</resources>
