+++ app/src/main/res/values/config.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <item name="hasTwoPanes" type="bool">false</item>
+</resources>
